return if deleted vfolder's parent is not dir

+++ src/ai/backend/storage/vfs/__init__.py
@@ -82,6 +82,9 @@ class BaseVolume(AbstractVolume):
         # remove intermediate prefix directories if they become empty
         from aiofiles import os as aiofile_os
 
+        if not path.parent.is_dir():
+            return
+
         if not os.listdir(path.parent):
             await aiofile_os.rmdir(path.parent)
